server/requirements.txt
fastapi
uvicorn
pydantic[email]
pydantic-settings
sqlalchemy
psycopg[binary]
python-jose
bcrypt>=4.0.0
passlib[bcrypt]
requests == 2.32.5
Pillow>=10.0.0
python-multipart
boto3>=1.34.0
botocore>=1.34.0
aws-requests-auth>=0.4.3
pytest
httpx